+++ include/BDSComponentFactory.hh
@@ -224,6 +224,7 @@ private:
   BDSAcceleratorComponent* CreateLaser();
   BDSAcceleratorComponent* CreateScreen();
   BDSAcceleratorComponent* CreateTransform3D();
+  BDSAcceleratorComponent* CreateWireScanner();
   BDSAcceleratorComponent* CreateRMatrix();
   BDSAcceleratorComponent* CreateThinRMatrix(G4double angleIn,
 					     const BDSMagnetStrength* stIn,